# Conflicts:
#	lib/Epub/Epub/Section.cpp
#	lib/Epub/Epub/css/CssParser.cpp
#	lib/Epub/Epub/parsers/ChapterHtmlSlimParser.cpp
#	lib/Epub/Epub/parsers/ChapterHtmlSlimParser.h
#	lib/I18n/translations/catalan.yaml
#	lib/I18n/translations/spanish.yaml
#	lib/I18n/translations/swedish.yaml
#	lib/I18n/translations/valencian.yaml
#	lib/Typesetter/Typesetter/ParsedText.cpp
#	lib/Typesetter/Typesetter/ParsedText.h
#	lib/Typesetter/Typesetter/blocks/BlockStyle.h
#	src/activities/reader/TxtReaderActivity.cpp

+++ src/activities/reader/ReaderUtils.h
@@ -9,10 +9,10 @@
 
 namespace ReaderUtils {
 
-constexpr unsigned long GO_HOME_MS = 1000;
-constexpr unsigned long SKIP_HOLD_MS = 700;
-constexpr unsigned long BOOKMARK_HOLD_MS = 400;
-constexpr unsigned long BOOKMARK_MESSAGE_DURATION_MS = 2500;
+constexpr uint32_t GO_HOME_MS = 1000;
+constexpr uint32_t SKIP_HOLD_MS = 700;
+constexpr uint32_t BOOKMARK_HOLD_MS = 400;
+constexpr uint32_t BOOKMARK_MESSAGE_DURATION_MS = 2500;
 
 inline void applyOrientation(GfxRenderer& renderer, const uint8_t orientation) {
   switch (orientation) {
